added react framework router links, renamed pages, added additional files for setup
