+++ frontend/src/components/shared/Footer.jsx
