+++ frontend/src/main.jsx
@@ -1,10 +1,9 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import '../src/pages/home.jsx'
+import React from 'react'
+import ReactDOM from 'react-dom/client'
 import App from '../src/App.jsx'
 
-createRoot(document.getElementById('root')).render(
-  <StrictMode>
+ReactDOM.createRoot(document.getElementById('root')).render(
+  <React.StrictMode>
     <App />
-  </StrictMode>
+  </React.StrictMode>
 )
